Move type definitions to separate file
