+++ karma.conf.js
@@ -9,7 +9,7 @@ module.exports = function(config) {
     frameworks: ['mocha'],
 
     preprocessors: {
-      'lib/contacthub.js': ['webpack', 'sourcemap'],
+      'src/contacthub.js': ['webpack', 'sourcemap'],
       'test/contacthub.js': ['webpack', 'sourcemap']
     },
 
@@ -17,7 +17,7 @@ module.exports = function(config) {
 
     files: [
       'test/queue.js',
-      'lib/contacthub.js',
+      'src/contacthub.js',
       'test/contacthub.js'
     ],
